+++ docs/tutorial_en.md
@@ -179,8 +179,16 @@ pip install pipenv
 pipenv install --dev
 ```
 
+Start update:
+
 ```python
 pipenv run dev
+```
+
+Start service:
+
+```python
+pipenv run service
 ```
 
 ### Method 3: GUI Software
