+++ tkinter_ui/tkinter_ui.py
@@ -7,7 +7,7 @@ from tkinter import messagebox
 from PIL import Image, ImageTk
 from utils.config import config
 from utils.tools import resource_path
-from main import UpdateSource, run_service
+from main import UpdateSource
 import asyncio
 import threading
 import webbrowser
@@ -32,6 +32,7 @@ if not ("TCL_LIBRARY" in environ and "TK_LIBRARY" in environ):
         tk_path = Path(base_prefix) / tk_dir
         environ["TCL_LIBRARY"] = str(next(tk_path.glob("tcl8.*")))
         environ["TK_LIBRARY"] = str(next(tk_path.glob("tk8.*")))
+
 
 class TkinterUI:
     def __init__(self, root):
@@ -126,15 +127,10 @@ class TkinterUI:
     def on_run_update(self):
         loop = asyncio.new_event_loop()
 
-        async def run_service_async():
-            loop.run_in_executor(None, run_service)
-
         def run_loop():
             asyncio.set_event_loop(loop)
             loop.run_until_complete(self.run_update())
 
-        if config.open_service:
-            asyncio.run(run_service_async())
         self.thread = threading.Thread(target=run_loop, daemon=True)
         self.thread.start()
